requirements.txt
# Core
django>=4.2,<5.2
djangorestframework>=3.14
django-istat-italian-places  # la tua app da PyPI

# DB + URL parsing
psycopg[binary]>=3.1
dj-database-url>=2.1

# Server + static files
gunicorn>=21
whitenoise[brotli]>=6.6
